Swap main.tsx and bootstrap.tsx content to align with module federation conventions
